Server Worker Module
Vite makes this super nice :)
https://stephendoddtech.com/blog/game-design/javascript-web-worker-set-interval-game-loop (Haven't read this yet, it was in my tab bar though)
https://vitejs.dev/guide/features.html#import-with-query-suffixes (sick!!! this helps out so much with the compilation step!)

I think I'm going to make myself allow myself (yes?) to rely on Vite in this case here. Getting the development of this set up is more important this time, than building it all from scratch. Of course, I still want to understand how it works under the hood though, I don't like magic code (I want to start using Svelte, but it isn't clicking for me how a lot of it works yet. At least, how it manages to do what it does. I know why it works, but I want to know how it works [is that the same thing? idk])

Tried setting up Svelte here, but it just isn't too easy to embed directly into the existing project. At least I don't know the best way to do that yet.

Just found this, not sure if it's what I'd like, but it kind of sounds like it.
https://github.com/DirtyHairy/worker-rpc