Interval-ed Game Loop
Got it working! The loop is instantiated in the worker, and the game is only rendered when a new tick is passed from the server. Now they are theoretically separate tasks which can also run at different speeds. The rendering itself should be as fast as the game should be viewed, but the ticks don't have to be as fast. Maybe they are faster? I can't remember. But the polling from server to client doesn't have to be as fast as the client rendering, for sure. The polling can be a different speed, and the server can probably run pretty fast, to calculate things properly.